// src/gen/coinbase/staking/v1alpha1/workflow.pb.ts
type Absent<T, K extends keyof T> = { [k in Exclude<keyof T, K>]?: undefined };
type OneOf<T> =
  | { [k in keyof T]?: undefined }
  | (
    keyof T extends infer K ?
      (K extends string & keyof T ? { [k in K]: T[K] } & Absent<T, K>
        : never)
    : never);
// ...
export enum WorkflowState {
  STATE_UNSPECIFIED = "STATE_UNSPECIFIED",
  STATE_IN_PROGRESS = "STATE_IN_PROGRESS",
  STATE_WAITING_FOR_SIGNING = "STATE_WAITING_FOR_SIGNING",
  STATE_COMPLETED = "STATE_COMPLETED",
  STATE_FAILED = "STATE_FAILED",
  STATE_CANCELING = "STATE_CANCELING",
  STATE_CANCELED = "STATE_CANCELED",
  STATE_CANCEL_FAILED = "STATE_CANCEL_FAILED",
  STATE_FAILED_REFRESHABLE = "STATE_FAILED_REFRESHABLE",
  STATE_WAITING_FOR_EXT_BROADCAST = "STATE_WAITING_FOR_EXT_BROADCAST",
}

export type TxStepOutput = {
  unsignedTx?: string
  signedTx?: string
  txHash?: string
  state?: TxStepOutputState
  errorMessage?: string
}

export type WaitStepOutput = {
  start?: string
  current?: string
  target?: string
  unit?: WaitStepOutputWaitUnit
  state?: WaitStepOutputState
}


type BaseWorkflowStep = {
  name?: string
}

export type WorkflowStep = BaseWorkflowStep
  & OneOf<{ txStepOutput: TxStepOutput; waitStepOutput: WaitStepOutput }>


type BaseWorkflow = {
  name?: string
  action?: string
  state?: WorkflowState
  currentStepId?: number
  steps?: WorkflowStep[]
  createTime?: GoogleProtobufTimestamp.Timestamp
  updateTime?: GoogleProtobufTimestamp.Timestamp
  skipBroadcast?: boolean
  completeTime?: GoogleProtobufTimestamp.Timestamp
}

export type Workflow = BaseWorkflow
  & OneOf<{ polygonStakingParameters: CoinbaseStakingV1alpha1Polygon.PolygonStakingParameters; solanaStakingParameters: CoinbaseStakingV1alpha1Solana.SolanaStakingParameters; ethereumKilnStakingParameters: CoinbaseStakingV1alpha1Ethereum_kiln.EthereumKilnStakingParameters }>

export type CreateWorkflowRequest = {
  parent?: string
  workflow?: Workflow
}